refactor(decompression): add animated unpacking types header
