+++ includes/acl/decompression/impl/steps/animated_unpack_types.h
@@ -1,0 +1,77 @@
+#pragma once
+
+////////////////////////////////////////////////////////////////////////////////
+// The MIT License (MIT)
+//
+// Copyright (c) 2025 Nicholas Frechette & Animation Compression Library contributors
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+////////////////////////////////////////////////////////////////////////////////
+
+#include "acl/version.h"
+#include "acl/core/impl/compiler_utils.h"
+
+#include <rtm/math.h>
+
+#include <cstdint>
+
+#define ACL_IMPL_USE_ANIMATED_PREFETCH
+
+ACL_IMPL_FILE_PRAGMA_PUSH
+
+namespace acl
+{
+	ACL_IMPL_VERSION_NAMESPACE_BEGIN
+
+	namespace acl_impl
+	{
+		struct clip_animated_sampling_context_v0
+		{
+			// Data is ordered in groups of 4 animated sub-tracks (e.g rot0, rot1, rot2, rot3)
+			// Order depends on animated track order. If we have 6 animated rotation tracks before the first animated
+			// translation track, we'll have 8 animated rotation sub-tracks followed by 4 animated translation sub-tracks.
+			// Once we reach the end, there is no extra padding. The last group might be less than 4 sub-tracks.
+			// This is because we always process 4 animated sub-tracks at a time and cache the results.
+
+			const uint8_t* clip_range_data;				// Range information of the current sub-track in the clip
+		};
+
+		struct segment_animated_sampling_context_v0
+		{
+			// Data is ordered in groups of 4 animated sub-tracks (e.g rot0, rot1, rot2, rot3)
+			// Order depends on animated track order. If we have 6 animated rotation tracks before the first animated
+			// translation track, we'll have 8 animated rotation sub-tracks followed by 4 animated translation sub-tracks.
+			// Once we reach the end, there is no extra padding. The last group might be less than 4 sub-tracks.
+			// This is because we always process 4 animated sub-tracks at a time and cache the results.
+
+			const uint8_t* format_per_track_data;		// Metadata of the current sub-track
+			const uint8_t* segment_range_data;			// Range information (or constant sample if bit rate is 0) of the current sub-track in this segment
+
+			// For the animated samples, constant bit rate sub-tracks (with a bit rate of 0) do not contain samples.
+			// As such, their group will not contain 4 sub-tracks.
+
+			const uint8_t* animated_track_data;			// Base of animated sample data, constant and doesn't change after init
+			uint32_t animated_track_data_bit_offset;	// Bit offset of the current animated sub-track
+		};
+	}
+
+	ACL_IMPL_VERSION_NAMESPACE_END
+}
+
+ACL_IMPL_FILE_PRAGMA_POP
